refactor. move functions out of DownloadButton into Row. Move buttons into their own class
